# backend/api/openapi/rime-v1.yaml
openapi: 3.1.0
info:
  title: Rime Native API
  version: 1.0.0
  description: Native API for catalog search and playback resolution.
servers:
  - url: /api/v1
security:
  - cookieAuth: []
  - bearerAuth: []
paths:
...
  /auth/setup:
    post:
      operationId: setupInitialAdmin
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SetupRequest' }
      responses:
        '201':
          description: Initial administrator created and signed in
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '400': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
  /auth/login:
    post:
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200':
          description: Signed in user
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '401': { $ref: '#/components/responses/Problem' }
        '429': { $ref: '#/components/responses/Problem' }
  /auth/token/setup:
    post:
      operationId: setupInitialAdminToken
      security: []
      parameters:
        - $ref: '#/components/parameters/NativeClient'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SetupRequest' }
      responses:
        '201':
          description: Initial administrator and native client session created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenSession' }
        '400': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
  /auth/token:
    post:
      operationId: createTokenSession
      security: []
      parameters:
        - $ref: '#/components/parameters/NativeClient'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200':
          description: Native client session created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenSession' }
        '401': { $ref: '#/components/responses/Problem' }
        '429': { $ref: '#/components/responses/Problem' }
...
  /me/password:
    patch:
      operationId: changeOwnPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [currentPassword, newPassword]
              properties:
                currentPassword: { type: string, format: password }
                newPassword: { type: string, format: password, minLength: 8, maxLength: 128 }
      responses:
        '204': { description: Password changed; all sessions revoked }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
  /me/playback-history:
    get:
      operationId: listOwnPlaybackHistory
      parameters:
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 50, default: 12 }
      responses:
        '200':
          description: 当前用户最近开始播放的歌曲；同一首歌只返回最新记录，不包含播放进度
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PlaybackHistoryPage' }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
  /me/playlists:
    get:
      operationId: listOwnPlaylists
      parameters:
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 50, default: 10 }
        - name: cursor
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Current user's playlists
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PlaylistPage' }
    post:
      operationId: createOwnPlaylist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, minLength: 1, maxLength: 80 }
      responses:
        '201':
          description: Playlist created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Playlist' }
  /me/playlists/{playlistId}:
    parameters:
      - $ref: '#/components/parameters/PlaylistId'
    get:
      operationId: getOwnPlaylist
      responses:
        '200':
          description: Playlist and ordered tracks
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PlaylistDetail' }
        '404': { $ref: '#/components/responses/Problem' }
    patch:
      operationId: renameOwnPlaylist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, minLength: 1, maxLength: 80 }
      responses:
        '200':
          description: Playlist renamed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Playlist' }
        '409': { $ref: '#/components/responses/Problem' }
    delete:
      operationId: deleteOwnPlaylist
      responses:
        '204': { description: Playlist deleted }
        '409': { $ref: '#/components/responses/Problem' }
  /me/playlists/{playlistId}/tracks:
    post:
      operationId: addTrackToOwnPlaylist
      parameters:
        - $ref: '#/components/parameters/PlaylistId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [trackId]
              properties:
                trackId: { type: string }
      responses:
        '204': { description: Track added }
        '409': { $ref: '#/components/responses/Problem' }
  /me/playlists/{playlistId}/tracks/{trackId}:
    delete:
      operationId: removeTrackFromOwnPlaylist
      parameters:
        - $ref: '#/components/parameters/PlaylistId'
        - $ref: '#/components/parameters/TrackId'
      responses:
        '204': { description: Track removed }
  /me/favorites/tracks/{trackId}:
    parameters:
      - $ref: '#/components/parameters/TrackId'
    get:
      operationId: getFavoriteStatus
      responses:
        '200':
          description: Favorite state for a track
          content:
            application/json:
              schema:
                type: object
                required: [favorite]
                properties:
                  favorite: { type: boolean }
    put:
      operationId: addFavorite
      responses:
        '204': { description: Track is a favorite }
    delete:
      operationId: removeFavorite
      responses:
        '204': { description: Track is not a favorite }
  /me/favorites/albums:
    get:
      operationId: listFavoriteAlbums
      parameters:
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 50, default: 12 }
      responses:
        '200':
          description: 当前用户直接喜欢且仍可播放的专辑
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AlbumPage' }
        '400': { $ref: '#/components/responses/Problem' }
  /me/favorites/albums/{albumId}:
    parameters:
      - name: albumId
        in: path
        required: true
        schema: { type: string }
    get:
      operationId: getFavoriteAlbumStatus
      responses:
        '200':
          description: 当前用户对专辑的直接喜欢状态
          content:
            application/json:
              schema:
                type: object
                required: [favorite]
                properties:
                  favorite: { type: boolean }
    put:
      operationId: addFavoriteAlbum
      responses:
        '204': { description: Album is a favorite }
        '404': { $ref: '#/components/responses/Problem' }
    delete:
      operationId: removeFavoriteAlbum
      responses:
        '204': { description: Album is not a favorite }
  /admin/users:
    get:
      operationId: listUsers
      responses:
        '200':
          description: All users; administrator only
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items: { type: array, items: { $ref: '#/components/schemas/User' } }
        '403': { $ref: '#/components/responses/Problem' }
    post:
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateUserRequest' }
      responses:
        '201':
          description: User created with a favorites playlist
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '403': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
  /admin/users/{userId}:
    patch:
      operationId: updateUser
      parameters:
        - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                displayName: { type: string, minLength: 1, maxLength: 40 }
                role: { type: string, enum: [admin, user] }
                disabled: { type: boolean }
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '403': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
  /admin/users/{userId}/password-reset:
    post:
      operationId: resetUserPassword
      parameters:
        - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password]
              properties:
                password: { type: string, format: password, minLength: 8, maxLength: 128 }
      responses:
        '204': { description: Temporary password set and sessions revoked }
        '403': { $ref: '#/components/responses/Problem' }
...
  /system/tasks/{taskId}/runs:
    post:
      operationId: runScheduledTask
      parameters:
        - name: taskId
          in: path
          required: true
          schema: { type: string }
      responses:
        '202':
          description: Scheduled task run started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduledTask'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
  /search:
    get:
      operationId: searchTracks
      parameters:
        - name: query
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 50, default: 20 }
        - name: cursor
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Ordered track search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchPage'
        '400':
          $ref: '#/components/responses/Problem'
  /albums:
    get:
      operationId: listAlbums
      parameters:
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 50, default: 24 }
        - name: cursor
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Albums ordered by normalized title
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlbumPage'
        '400':
          $ref: '#/components/responses/Problem'
  /albums/recent:
    get:
      operationId: listRecentAlbums
      parameters:
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 50, default: 12 }
        - name: cursor
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Albums ordered by the most recent media library ingestion time
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlbumPage'
        '400':
          $ref: '#/components/responses/Problem'
  /albums/{albumId}:
    get:
      operationId: getAlbumDetail
      parameters:
        - name: albumId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Album metadata and its available tracks
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlbumDetail'
        '404':
          $ref: '#/components/responses/Problem'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: rime_session
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    NativeClient:
      name: X-Rime-Client
      in: header
      required: true
      schema: { type: string, enum: [tauri] }
...
  responses:
    Problem:
      description: RFC 9457 problem details
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
    User:
      type: object
      required: [id, username, displayName, role, disabled, mustChangePassword, createdAt]
      properties:
        id: { type: string }
        username: { type: string }
        displayName: { type: string }
        role: { type: string, enum: [admin, user] }
        disabled: { type: boolean }
        mustChangePassword: { type: boolean }
        createdAt: { type: string, format: date-time }
        lastLoginAt: { type: string, format: date-time }
...
    SetupRequest:
      type: object
      required: [username, password]
      properties:
        username: { type: string, minLength: 3, maxLength: 32 }
        displayName: { type: string, maxLength: 40 }
        password: { type: string, format: password, minLength: 8, maxLength: 128 }
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username: { type: string }
        password: { type: string, format: password }
...