# backend/internal/transport/http/native/v1/handler.go
package v1

import (
	"encoding/json"
	"errors"
	"fmt"
	"io"
	"log/slog"
	"mime"
	"net/http"
	"net/url"
	"os"
	"path/filepath"
	"strconv"
	"strings"
	"time"

	"rime/backend/internal/artwork"
	"rime/backend/internal/browse"
	"rime/backend/internal/identity"
	"rime/backend/internal/lyrics"
	"rime/backend/internal/playback"
	"rime/backend/internal/playlists"
	"rime/backend/internal/search"
	"rime/backend/internal/tasks"
)

const (
	nativeClientHeader = "X-Rime-Client"
	nativeClientTauri  = "tauri"
)

type Handler struct {
	search    *search.Service
	browse    *browse.Service
	lyrics    *lyrics.Service
	playback  *playback.Service
	artwork   *artwork.Service
	tasks     *tasks.Service
	identity  *identity.Service
	playlists *playlists.Service
	logger    *slog.Logger
}

func New(searchService *search.Service, browseService *browse.Service, lyricsService *lyrics.Service, playbackService *playback.Service, artworkService *artwork.Service, taskService *tasks.Service, identityService *identity.Service, playlistService *playlists.Service, logger *slog.Logger) http.Handler {
	handler := &Handler{search: searchService, browse: browseService, lyrics: lyricsService, playback: playbackService, artwork: artworkService, tasks: taskService, identity: identityService, playlists: playlistService, logger: logger}
	mux := http.NewServeMux()
	mux.HandleFunc("GET /api/v1/auth/status", handler.authStatus)
	mux.HandleFunc("POST /api/v1/auth/setup", handler.setupAdmin)
	mux.HandleFunc("POST /api/v1/auth/login", handler.login)
	mux.HandleFunc("POST /api/v1/auth/token/setup", handler.setupAdminToken)
	mux.HandleFunc("POST /api/v1/auth/token", handler.loginToken)
	mux.HandleFunc("DELETE /api/v1/auth/session", handler.logout)
	mux.HandleFunc("GET /api/v1/me", handler.me)
	mux.HandleFunc("PATCH /api/v1/me/password", handler.changePassword)
	mux.HandleFunc("GET /api/v1/me/playback-history", handler.recentPlaybackTracks)
	mux.HandleFunc("GET /api/v1/me/playlists", handler.listPlaylists)
	mux.HandleFunc("POST /api/v1/me/playlists", handler.createPlaylist)
	mux.HandleFunc("GET /api/v1/me/playlists/{playlistID}", handler.getPlaylist)
	mux.HandleFunc("PATCH /api/v1/me/playlists/{playlistID}", handler.renamePlaylist)
	mux.HandleFunc("DELETE /api/v1/me/playlists/{playlistID}", handler.deletePlaylist)
	mux.HandleFunc("POST /api/v1/me/playlists/{playlistID}/tracks", handler.addPlaylistTrack)
	mux.HandleFunc("DELETE /api/v1/me/playlists/{playlistID}/tracks/{trackID}", handler.removePlaylistTrack)
	mux.HandleFunc("GET /api/v1/me/favorites/albums", handler.favoriteAlbums)
	mux.HandleFunc("GET /api/v1/me/favorites/albums/{albumID}", handler.favoriteAlbumStatus)
	mux.HandleFunc("PUT /api/v1/me/favorites/albums/{albumID}", handler.addFavoriteAlbum)
	mux.HandleFunc("DELETE /api/v1/me/favorites/albums/{albumID}", handler.removeFavoriteAlbum)
	mux.HandleFunc("GET /api/v1/me/favorites/tracks/{trackID}", handler.favoriteStatus)
	mux.HandleFunc("PUT /api/v1/me/favorites/tracks/{trackID}", handler.addFavorite)
	mux.HandleFunc("DELETE /api/v1/me/favorites/tracks/{trackID}", handler.removeFavorite)
	mux.HandleFunc("GET /api/v1/admin/users", handler.listUsers)
	mux.HandleFunc("POST /api/v1/admin/users", handler.createUser)
	mux.HandleFunc("PATCH /api/v1/admin/users/{userID}", handler.updateUser)
	mux.HandleFunc("POST /api/v1/admin/users/{userID}/password-reset", handler.resetPassword)
	mux.HandleFunc("GET /api/v1/system/info", handler.systemInfo)
	mux.HandleFunc("GET /api/v1/system/tasks", handler.listTasks)
	mux.HandleFunc("POST /api/v1/system/tasks/{taskID}/runs", handler.runTask)
	mux.HandleFunc("GET /api/v1/search", handler.searchTracks)
	mux.HandleFunc("GET /api/v1/albums", handler.albums)
	mux.HandleFunc("GET /api/v1/albums/recent", handler.recentAlbums)
	mux.HandleFunc("GET /api/v1/albums/{albumID}", handler.albumDetail)
	mux.HandleFunc("GET /api/v1/artists/{artistID}", handler.artistDetail)
	mux.HandleFunc("GET /api/v1/tracks/{trackID}/lyrics", handler.trackLyrics)
	mux.HandleFunc("POST /api/v1/playback/sessions", handler.createPlaybackSession)
	mux.HandleFunc("GET /api/v1/playback/sessions/{sessionID}/stream", handler.stream)
	mux.HandleFunc("HEAD /api/v1/playback/sessions/{sessionID}/stream", handler.stream)
	mux.HandleFunc("POST /api/v1/playback/sessions/{sessionID}/events", handler.recordPlaybackEvent)
	mux.HandleFunc("DELETE /api/v1/playback/sessions/{sessionID}", handler.deletePlaybackSession)
	mux.HandleFunc("GET /api/v1/artworks/{artworkID}", handler.serveArtwork)
	mux.HandleFunc("HEAD /api/v1/artworks/{artworkID}", handler.serveArtwork)
	mux.HandleFunc("GET /healthz", handler.health)
	return requestMiddleware(logger, authenticationMiddleware(handler, mux))
}

func (h *Handler) serveArtwork(w http.ResponseWriter, r *http.Request) {
	size, err := optionalInt(r.URL.Query().Get("size"))
	if err != nil {
		writeProblem(w, r, http.StatusBadRequest, "invalid_artwork_size", "Invalid artwork size", "Artwork size must be an integer.")
		return
	}
	resource, err := h.artwork.Open(r.Context(), r.PathValue("artworkID"), size)
	if err != nil {
		switch {
		case errors.Is(err, artwork.ErrInvalidSize):
			writeProblem(w, r, http.StatusBadRequest, "invalid_artwork_size", "Invalid artwork size", "Supported sizes are 128, 256, 512, and 1024 pixels.")
		case errors.Is(err, artwork.ErrNotFound):
			writeProblem(w, r, http.StatusNotFound, "artwork_not_found", "Artwork not found", "The artwork is missing or its cache must be rebuilt.")
		default:
			h.logger.Error("serve artwork", "artwork_id", r.PathValue("artworkID"), "error", err)
			writeProblem(w, r, http.StatusInternalServerError, "internal_error", "Internal error", "The artwork could not be loaded.")
		}
		return
	}
	defer resource.File.Close()
	w.Header().Set("Cache-Control", "private, max-age=31536000, immutable")
	w.Header().Set("Content-Type", resource.ContentType)
	w.Header().Set("ETag", fmt.Sprintf("\"%s-%d\"", r.PathValue("artworkID"), size))
	http.ServeContent(w, r, resource.Name, resource.ModifiedAt, resource.File)
}

func (h *Handler) systemInfo(w http.ResponseWriter, _ *http.Request) {
	writeJSON(w, http.StatusOK, map[string]any{
		"name":       "Rime",
		"apiVersion": "v1",
		"capabilities": []string{
			"auth.bearer.v1",
			"search.tracks.v1",
			"browse.albums.v1",
			"browse.recent-albums.v1",
			"lyrics.timed.v1",
			"playback.direct.v1",
			"playback.events.v1",
			"playback.history.v1",
			"favorites.albums.v1",
			"system.tasks.v1",
		},
	})
}

func (h *Handler) trackLyrics(w http.ResponseWriter, r *http.Request) {
	document, err := h.lyrics.Get(r.Context(), r.PathValue("trackID"))
	if err != nil {
		if errors.Is(err, lyrics.ErrNotFound) {
			writeProblem(w, r, http.StatusNotFound, "lyrics_not_found", "Lyrics not found", "No lyrics are available for this track.")
			return
		}
		h.logger.Error("get track lyrics", "track_id", r.PathValue("trackID"), "error", err)
		writeProblem(w, r, http.StatusInternalServerError, "internal_error", "Internal error", "The lyrics could not be loaded.")
		return
	}
	writeJSON(w, http.StatusOK, document)
}

// albums 返回按标题排序的全部可播放专辑游标页。
// 参数 w 写入 JSON 响应，r 提供 limit、cursor 查询参数及请求上下文；非法分页参数
// 返回 400，游标不能由本接口续用时同样返回 400，其他读取错误返回 500。
func (h *Handler) albums(w http.ResponseWriter, r *http.Request) {
	limit, err := optionalInt(r.URL.Query().Get("limit"))
	if err != nil {
		writeProblem(w, r, http.StatusBadRequest, "invalid_limit", "Invalid limit", "Limit must be an integer.")
		return
	}
	page, err := h.browse.Albums(r.Context(), limit, r.URL.Query().Get("cursor"))
	if err != nil {
		if errors.Is(err, browse.ErrInvalidLimit) {
			writeProblem(w, r, http.StatusBadRequest, "invalid_limit", "Invalid limit", err.Error())
			return
		}
		if errors.Is(err, browse.ErrInvalidCursor) {
			writeProblem(w, r, http.StatusBadRequest, "invalid_cursor", "Invalid cursor", "Cursor must be a value returned by this endpoint.")
			return
		}
		h.logger.Error("list albums", "error", err)
		writeProblem(w, r, http.StatusInternalServerError, "internal_error", "Internal error", "Albums could not be loaded.")
		return
	}
	writeJSON(w, http.StatusOK, page)
}

func (h *Handler) recentAlbums(w http.ResponseWriter, r *http.Request) {
	limit, err := optionalInt(r.URL.Query().Get("limit"))
	if err != nil {
		writeProblem(w, r, http.StatusBadRequest, "invalid_limit", "Invalid limit", "Limit must be an integer.")
		return
	}
	page, err := h.browse.RecentAlbums(r.Context(), limit, r.URL.Query().Get("cursor"))
	if err != nil {
		if errors.Is(err, browse.ErrInvalidLimit) {
			writeProblem(w, r, http.StatusBadRequest, "invalid_limit", "Invalid limit", err.Error())
			return
		}
		if errors.Is(err, browse.ErrInvalidCursor) {
			writeProblem(w, r, http.StatusBadRequest, "invalid_cursor", "Invalid cursor", "Cursor must be a value returned by this endpoint.")
			return
		}
		h.logger.Error("list recent albums", "error", err)
		writeProblem(w, r, http.StatusInternalServerError, "internal_error", "Internal error", "Recent albums could not be loaded.")
		return
	}
	writeJSON(w, http.StatusOK, page)
}

// albumDetail 返回一个专辑及其全部可播放曲目。
// 参数 w 写入 HTTP 响应，r 提供 albumID 路径参数及请求上下文。
// 专辑不存在或没有可播放曲目时返回 404，其他存储错误返回 500。
func (h *Handler) albumDetail(w http.ResponseWriter, r *http.Request) {
	detail, err := h.browse.AlbumDetail(r.Context(), r.PathValue("albumID"))
	if err != nil {
		if errors.Is(err, browse.ErrAlbumNotFound) {
			writeProblem(w, r, http.StatusNotFound, "album_not_found", "Album not found", "The requested album is unavailable.")
			return
		}
		h.logger.Error("get album detail", "album_id", r.PathValue("albumID"), "error", err)
		writeProblem(w, r, http.StatusInternalServerError, "internal_error", "Internal error", "The album could not be loaded.")
		return
	}
	writeJSON(w, http.StatusOK, detail)
}

// artistDetail 返回一个歌手及其参与的可播放专辑批次。
// 参数 w 写入 HTTP 响应，r 提供 artistID、limit 和 cursor；歌手不存在或没有可播放
// 曲目时返回 404，无效分页参数返回 400，其他存储错误返回 500。
func (h *Handler) artistDetail(w http.ResponseWriter, r *http.Request) {
	limit, err := optionalInt(r.URL.Query().Get("limit"))
	if err != nil {
		writeProblem(w, r, http.StatusBadRequest, "invalid_limit", "Invalid limit", "Limit must be an integer.")
		return
	}
	detail, err := h.browse.ArtistDetail(r.Context(), r.PathValue("artistID"), limit, r.URL.Query().Get("cursor"))
	if err != nil {
		if errors.Is(err, browse.ErrInvalidLimit) {
			writeProblem(w, r, http.StatusBadRequest, "invalid_limit", "Invalid limit", err.Error())
			return
		}
		if errors.Is(err, browse.ErrInvalidCursor) {
			writeProblem(w, r, http.StatusBadRequest, "invalid_cursor", "Invalid cursor", "Cursor must be a value returned by this endpoint.")
			return
		}
		if errors.Is(err, browse.ErrArtistNotFound) {
			writeProblem(w, r, http.StatusNotFound, "artist_not_found", "Artist not found", "The requested artist is unavailable.")
			return
		}
		h.logger.Error("get artist detail", "artist_id", r.PathValue("artistID"), "error", err)
		writeProblem(w, r, http.StatusInternalServerError, "internal_error", "Internal error", "The artist could not be loaded.")
		return
	}
	writeJSON(w, http.StatusOK, detail)
}

func (h *Handler) listTasks(w http.ResponseWriter, r *http.Request) {
	if !h.requireAdmin(w, r) {
		return
	}
	items, err := h.tasks.List(r.Context())
	if err != nil {
		h.logger.Error("list scheduled tasks", "error", err)
		writeProblem(w, r, http.StatusInternalServerError, "internal_error", "Internal error", "The scheduled tasks could not be loaded.")
		return
	}
	writeJSON(w, http.StatusOK, map[string]any{"items": items})
}

func (h *Handler) runTask(w http.ResponseWriter, r *http.Request) {
	if !h.requireAdmin(w, r) {
		return
	}
	task, err := h.tasks.RunNow(r.Context(), r.PathValue("taskID"))
	if err != nil {
		switch {
		case errors.Is(err, tasks.ErrNotFound):
			writeProblem(w, r, http.StatusNotFound, "scheduled_task_not_found", "Scheduled task not found", "The requested scheduled task does not exist.")
		case errors.Is(err, tasks.ErrAlreadyRunning):
			writeProblem(w, r, http.StatusConflict, "scheduled_task_running", "Scheduled task is running", "The requested scheduled task is already running.")
		default:
			h.logger.Error("run scheduled task", "task_id", r.PathValue("taskID"), "error", err)
			writeProblem(w, r, http.StatusInternalServerError, "internal_error", "Internal error", "The scheduled task could not be started.")
		}
		return
	}
	writeJSON(w, http.StatusAccepted, task)
}

func (h *Handler) health(w http.ResponseWriter, _ *http.Request) {
	writeJSON(w, http.StatusOK, map[string]string{"status": "ok"})
}

func (h *Handler) searchTracks(w http.ResponseWriter, r *http.Request) {
	limit, err := optionalInt(r.URL.Query().Get("limit"))
	if err != nil {
		writeProblem(w, r, http.StatusBadRequest, "invalid_limit", "Invalid limit", err.Error())
		return
	}
	page, err := h.search.Tracks(r.Context(), r.URL.Query().Get("query"), limit, r.URL.Query().Get("cursor"))
	if err != nil {
		writeProblem(w, r, http.StatusBadRequest, "invalid_search", "Invalid search", err.Error())
		return
	}
	writeJSON(w, http.StatusOK, page)
}

// recentPlaybackTracks 返回当前登录用户的继续聆听歌曲。
//
// 参数 w 用于写入 JSON 响应，r 提供可选 limit 查询参数和身份上下文。历史记录只在歌曲
// 开始播放时持久化，服务端最多保留最近 300 条；本接口仅允许读取其中 1 至 50 条，且不返回
// 播放进度。非法 limit 返回 400，读取失败返回 500。
func (h *Handler) recentPlaybackTracks(w http.ResponseWriter, r *http.Request) {
	limitValue := r.URL.Query().Get("limit")
	limit, err := optionalInt(limitValue)
	if err != nil {
		writeProblem(w, r, http.StatusBadRequest, "invalid_limit", "Invalid limit", "Limit must be an integer.")
		return
	}
	if limitValue != "" && (limit < 1 || limit > 50) {
		writeProblem(w, r, http.StatusBadRequest, "invalid_limit", "Invalid limit", "Limit must be between 1 and 50.")
		return
	}
	page, err := h.playback.RecentTracks(r.Context(), currentUser(r).ID, limit)
	if err != nil {
		h.logger.Error("list playback history", "user_id", currentUser(r).ID, "error", err)
		writeProblem(w, r, http.StatusInternalServerError, "internal_error", "Internal error", "Playback history could not be loaded.")
		return
	}
	writeJSON(w, http.StatusOK, page)
}

func (h *Handler) createPlaybackSession(w http.ResponseWriter, r *http.Request) {
	var request playback.CreateRequest
	if err := decodeJSON(r, &request); err != nil {
		writeProblem(w, r, http.StatusBadRequest, "invalid_request", "Invalid request", err.Error())
		return
	}
	session, err := h.playback.Create(r.Context(), currentUser(r).ID, request)
	if err != nil {
		switch {
		case errors.Is(err, playback.ErrTrackNotFound):
			writeProblem(w, r, http.StatusNotFound, "track_not_found", "Track not found", "The requested track is unavailable.")
		case errors.Is(err, playback.ErrUnsupportedFormat):
			writeProblem(w, r, http.StatusConflict, "playback_format_unsupported", "Playback format unsupported", "No direct-play source matches this player.")
		default:
			h.logger.Error("create playback session", "error", err)
			writeProblem(w, r, http.StatusInternalServerError, "internal_error", "Internal error", "The playback session could not be created.")
		}
		return
	}
	w.Header().Set("Location", "/api/v1/playback/sessions/"+session.SessionID)
	writeJSON(w, http.StatusCreated, session)
}

func (h *Handler) stream(w http.ResponseWriter, r *http.Request) {
	track, mediaFile, err := h.playback.Stream(r.Context(), r.PathValue("sessionID"))
	if err != nil {
		if errors.Is(err, playback.ErrSessionNotFound) {
			writeProblem(w, r, http.StatusNotFound, "session_not_found", "Playback session not found", "The playback session is missing or expired.")
			return
		}
		h.logger.Error("resolve playback stream", "error", err)
		writeProblem(w, r, http.StatusInternalServerError, "internal_error", "Internal error", "The media source could not be resolved.")
		return
	}
	file, err := os.Open(mediaFile.Path)
	if err != nil {
		if errors.Is(err, os.ErrNotExist) {
			writeProblem(w, r, http.StatusGone, "media_missing", "Media missing", "The indexed media file is no longer available.")
			return
		}
		h.logger.Error("open media stream", "error", err)
		writeProblem(w, r, http.StatusInternalServerError, "internal_error", "Internal error", "The media file could not be opened.")
		return
	}
	defer file.Close()

	name := safeFilename(track.Title, filepath.Ext(mediaFile.Path))
	w.Header().Set("Accept-Ranges", "bytes")
	w.Header().Set("Cache-Control", "private, max-age=0, must-revalidate")
	w.Header().Set("Content-Disposition", mime.FormatMediaType("inline", map[string]string{"filename": name}))
	w.Header().Set("Content-Type", mediaFile.ContentType)
	w.Header().Set("ETag", fmt.Sprintf("\"%s\"", mediaFile.ContentVersion))
	http.ServeContent(w, r, name, time.UnixMilli(mediaFile.ModifiedUnixMs), file)
}

func (h *Handler) recordPlaybackEvent(w http.ResponseWriter, r *http.Request) {
	var event playback.Event
	if err := decodeJSON(r, &event); err != nil {
		writeProblem(w, r, http.StatusBadRequest, "invalid_request", "Invalid request", err.Error())
		return
	}
	if err := h.playback.Record(r.Context(), currentUser(r).ID, r.PathValue("sessionID"), event); err != nil {
		writeProblem(w, r, http.StatusBadRequest, "invalid_playback_event", "Invalid playback event", err.Error())
		return
	}
	w.WriteHeader(http.StatusNoContent)
}

func (h *Handler) deletePlaybackSession(w http.ResponseWriter, r *http.Request) {
	if err := h.playback.Delete(r.Context(), currentUser(r).ID, r.PathValue("sessionID")); err != nil {
		h.logger.Error("delete playback session", "error", err)
		writeProblem(w, r, http.StatusInternalServerError, "internal_error", "Internal error", "The playback session could not be deleted.")
		return
	}
	w.WriteHeader(http.StatusNoContent)
}

type problem struct {
	Type     string `json:"type"`
	Title    string `json:"title"`
	Status   int    `json:"status"`
	Detail   string `json:"detail"`
	Instance string `json:"instance"`
	Code     string `json:"code"`
	TraceID  string `json:"traceId,omitempty"`
}

func writeProblem(w http.ResponseWriter, r *http.Request, status int, code, title, detail string) {
	w.Header().Set("Content-Type", "application/problem+json")
	w.WriteHeader(status)
	_ = json.NewEncoder(w).Encode(problem{
		Type:     "about:blank",
		Title:    title,
		Status:   status,
		Detail:   detail,
		Instance: r.URL.Path,
		Code:     code,
		TraceID:  w.Header().Get("X-Request-ID"),
	})
}

func writeJSON(w http.ResponseWriter, status int, value any) {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(status)
	_ = json.NewEncoder(w).Encode(value)
}

func decodeJSON(r *http.Request, target any) error {
	decoder := json.NewDecoder(io.LimitReader(r.Body, 64<<10))
	decoder.DisallowUnknownFields()
	if err := decoder.Decode(target); err != nil {
		return err
	}
	if err := decoder.Decode(&struct{}{}); !errors.Is(err, io.EOF) {
		return fmt.Errorf("request body must contain one JSON value")
	}
	return nil
}

func optionalInt(value string) (int, error) {
	if value == "" {
		return 0, nil
	}
	return strconv.Atoi(value)
}

func safeFilename(title, extension string) string {
	replacer := strings.NewReplacer("/", "_", "\\", "_", "\r", "", "\n", "")
	return replacer.Replace(title) + extension
}

func requestMiddleware(logger *slog.Logger, next http.Handler) http.Handler {
	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		started := time.Now()
		requestID := r.Header.Get("X-Request-ID")
		if requestID == "" {
			requestID = strconv.FormatInt(started.UnixNano(), 36)
		}
		w.Header().Set("X-Request-ID", requestID)
		w.Header().Set("Rime-API-Version", "v1")
		w.Header().Set("X-Content-Type-Options", "nosniff")
		if isUnsafeMethod(r.Method) && !hasBearerToken(r) && !isNativeTokenRequest(r) && !isSameOriginRequest(r) {
			writeProblem(w, r, http.StatusForbidden, "cross_origin_request", "Request blocked", "State-changing requests must originate from this Rime server.")
			return
		}
		next.ServeHTTP(w, r)
		logger.Debug("http request", "method", r.Method, "path", r.URL.Path, "duration", time.Since(started), "request_id", requestID)
	})
}

func isNativeTokenRequest(r *http.Request) bool {
	if r.Method != http.MethodPost || !strings.EqualFold(strings.TrimSpace(r.Header.Get(nativeClientHeader)), nativeClientTauri) {
		return false
	}
	return r.URL.Path == "/api/v1/auth/token" || r.URL.Path == "/api/v1/auth/token/setup"
}

func isUnsafeMethod(method string) bool {
	return method == http.MethodPost || method == http.MethodPut || method == http.MethodPatch || method == http.MethodDelete
}

func isSameOriginRequest(r *http.Request) bool {
	if strings.EqualFold(r.Header.Get("Sec-Fetch-Site"), "cross-site") {
		return false
	}
	origin := r.Header.Get("Origin")
	if origin == "" {
		return true
	}
	parsed, err := url.Parse(origin)
	if err != nil {
		return false
	}
	scheme := "http"
	if requestIsHTTPS(r) {
		scheme = "https"
	}
	return strings.EqualFold(parsed.Scheme, scheme) && strings.EqualFold(parsed.Host, r.Host)
}